+++ Sources/Parley/Data/Remotes/ParleyRemote.swift
@@ -169,10 +169,10 @@ final class ParleyRemote {
     func execute<T: Codable>(
         _ method: ParleyHTTPRequestMethod = .post,
         path: String,
-        imageData: Data,
+        data: Data,
         name: String,
         fileName: String,
-        imageType: ParleyImageType,
+        type: ParleyMediaType,
         result: @escaping (Result<T, Error>) -> Void
     ) {
         debugPrint("ParleyRemote.execute:: \(method) \(getUrl(path))")
@@ -181,8 +181,8 @@ final class ParleyRemote {
         multipartFormData.add(
             key: "media",
             fileName: fileName,
-            fileMimeType: imageType.rawValue,
-            fileData: imageData
+            fileMimeType: type.rawValue,
+            fileData: data
         )
         var headers = createHeaders()
         headers[HTTPHeaders.contentType.rawValue] = multipartFormData.httpContentTypeHeaderValue
@@ -232,13 +232,13 @@ final class ParleyRemote {
         }
     }
 
-    // MARK: - Image
+    // MARK: - Media
 
     func execute(
         _ method: ParleyHTTPRequestMethod,
         path: String,
-        type: ParleyImageType,
-        result: @escaping (Result<ParleyImageNetworkModel, Error>) -> Void
+        type: ParleyMediaType,
+        result: @escaping (Result<Data, Error>) -> Void
     ) {
         let url = getUrl(path)
         debugPrint("ParleyRemote.execute:: \(method) \(getUrl(path))")
@@ -256,7 +256,7 @@ final class ParleyRemote {
                         switch requestResult {
                         case .success(let response):
                             if let data = response.body {
-                                result(.success(ParleyImageNetworkModel(data: data, type: type)))
+                                result(.success(data))
                             }
                         case .failure(let error):
                             if let data = error.data, let apiError = Self.decodeBackendError(responseData: data) {