+++ Sources/Parley/Presentation/Models/FileDisplayModel.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-struct FileDisplayModel: Hashable {
-    
-    // TODO: Fill with display components
-    let location: URL
-}