+++ Sources/Parley/Views/CollectionViewCells/MessageCollectionViewCell/MessageCollectionViewCell.swift
@@ -20,8 +20,8 @@ final class MessageCollectionViewCell: UICollectionViewCell {
         }
     }
 
-    func render(_ message: Message, time: Date?, mediaLoader: MediaLoaderProtocol?) {
-        parleyMessageView.set(message: message, forcedTime: time, mediaLoader: mediaLoader)
+    func render(_ message: Message, time: Date?, mediaLoader: MediaLoaderProtocol?, shareManager: ShareManager?) {
+        parleyMessageView.set(message: message, forcedTime: time, mediaLoader: mediaLoader, shareManager: shareManager)
         setupAccessibilityOptions(for: message)
     }
 
@@ -65,11 +65,13 @@ final class MessageCollectionViewCell: UICollectionViewCell {
         totalHeight += appearance.balloonContentInsets?.top ?? 0
         totalHeight += appearance.balloonContentInsets?.bottom ?? 0
 
-        if message.hasMedium {
+        var showTopSeparator: Bool = true
+        if message.hasImage {
             totalHeight += 160
 
             totalHeight += appearance.imageInsets?.top ?? 0
             totalHeight += appearance.imageInsets?.bottom ?? 0
+            showTopSeparator = false
         } else if let name = message.agent?.name, appearance.name {
             totalHeight += appearance.nameInsets?.top ?? 0
             totalHeight += appearance.nameInsets?.bottom ?? 0
@@ -92,8 +94,60 @@ final class MessageCollectionViewCell: UICollectionViewCell {
                 font: appearance.messageTextViewAppearance.regularFont
             )
         }
+        
+        if message.hasFile {
+            if showTopSeparator {
+                // Top separator
+                totalHeight += appearance.separatorInset?.top ?? 0
+                totalHeight += appearance.separatorInset?.bottom ?? 0
+                totalHeight += 1
+            }
+            
+            totalHeight += appearance.fileContentInsets?.top ?? 0
+            totalHeight += appearance.fileContentInsets?.bottom ?? 0
+            totalHeight += max(
+                ParleyLocalizationKey.messageFileOpen.localized.height(withConstrainedWidth: contentWidth, font: appearance.buttonFont),
+                message.media?.displayFileName.height(withConstrainedWidth: contentWidth, font: appearance.fileNameFont) ?? 0
+            )
+            
+            // Bottom separator
+            totalHeight += appearance.separatorInset?.top ?? 0
+            totalHeight += appearance.separatorInset?.bottom ?? 0
+            totalHeight += 1
+            
+            showTopSeparator = false
+        }
 
-        if message.message != nil || message.title != nil || message.hasButtons || !message.hasMedium {
+        if message.hasButtons {
+            totalHeight += appearance.buttonsInsets?.top ?? 0
+            totalHeight += appearance.buttonsInsets?.bottom ?? 0
+            
+            var buttonWidth = balloonWidth
+            buttonWidth -= appearance.buttonsInsets?.left ?? 0
+            buttonWidth -= appearance.buttonsInsets?.right ?? 0
+            buttonWidth -= appearance.buttonInsets?.left ?? 0
+            buttonWidth -= appearance.buttonInsets?.right ?? 0
+            
+            if showTopSeparator {
+                totalHeight += 1 // Initial separator
+                totalHeight += appearance.separatorInset?.top ?? 0
+                totalHeight += appearance.separatorInset?.bottom ?? 0
+            }
+            message.buttons?.forEach({ (button: MessageButton) in
+                totalHeight += appearance.buttonInsets?.top ?? 0
+                totalHeight += appearance.buttonInsets?.bottom ?? 0
+                totalHeight += button.title.height(withConstrainedWidth: buttonWidth, font: appearance.buttonFont)
+                totalHeight += 1 // Button separator
+                totalHeight += appearance.separatorInset?.top ?? 0
+                totalHeight += appearance.separatorInset?.bottom ?? 0
+            })
+
+            if message.message == nil && message.title == nil && !message.hasImage {
+                totalHeight += appearance.balloonContentTextInsets?.top ?? 0
+            }
+        }
+        
+        if message.message != nil || message.title != nil || message.hasButtons || !message.hasImage {
             totalHeight += appearance.metaInsets?.top ?? 0
             totalHeight += appearance.metaInsets?.bottom ?? 0
 
@@ -101,30 +155,7 @@ final class MessageCollectionViewCell: UICollectionViewCell {
                 .height(withConstrainedWidth: contentWidth, font: appearance.timeFont)
         }
 
-        if message.hasButtons {
-            totalHeight += appearance.buttonsInsets?.top ?? 0
-            totalHeight += appearance.buttonsInsets?.bottom ?? 0
-
-            var buttonWidth = balloonWidth
-            buttonWidth -= appearance.buttonsInsets?.left ?? 0
-            buttonWidth -= appearance.buttonsInsets?.right ?? 0
-            buttonWidth -= appearance.buttonInsets?.left ?? 0
-            buttonWidth -= appearance.buttonInsets?.right ?? 0
-
-            totalHeight += 1 // Initial separator
-            message.buttons?.forEach({ (button: MessageButton) in
-                totalHeight += appearance.buttonInsets?.top ?? 0
-                totalHeight += appearance.buttonInsets?.bottom ?? 0
-                totalHeight += button.title.height(withConstrainedWidth: buttonWidth, font: appearance.buttonFont)
-                totalHeight += 1 // Button separator
-            })
-
-            if message.message == nil && message.title == nil && !message.hasMedium {
-                totalHeight += appearance.balloonContentTextInsets?.top ?? 0
-            }
-        }
-
-        if message.message != nil || message.title != nil || message.hasMedium {
+        if message.message != nil || message.title != nil || message.hasImage {
             totalHeight += appearance.balloonContentTextInsets?.top ?? 0
             totalHeight += appearance.balloonContentTextInsets?.bottom ?? 0
         }