+++ Sources/Parley/Views/ParleyComposeView/ParleyComposeViewDelegate.swift
@@ -7,6 +7,7 @@ protocol ParleyComposeViewDelegate: AnyObject {
 
     func send(_ message: String)
     func send(image: UIImage, with data: Data, url: URL)
+    func send(file url: URL)
 
     @available(iOS 14.0, *)
     func send(image: UIImage, data: Data, fileName: String, type: UTType)